revamping file system, adding in more configuration options, updating checksum calculation, atomic writer, and making sure the corruption is properly handled

+++ Source/Nomad.Core/Util/BufferHandles/SharedBufferHandle.cs
@@ -25,7 +25,7 @@ namespace Nomad.Core.Util.BufferHandles
         /// </summary>
         /// <param name="buffer"></param>
         /// <param name="length"></param>
-        public SharedBufferHandle(byte[] buffer, int length)
+        public SharedBufferHandle(byte[] buffer, long length)
             : base(buffer, length)
         {
         }
